chore: add backwards compatibility to packages
BG-50243
