+++ modules/bitgo/src/index.ts
@@ -7,7 +7,7 @@
 // Copyright 2022, BitGo, Inc.  All Rights Reserved.
 //
 import * as _ from 'lodash';
-import { common, CustomSigningFunction, SignedTransaction, tss } from '@bitgo/sdk-core';
+import { common } from '@bitgo/sdk-core';
 export * from '@bitgo/sdk-api';
 import * as utxolib from '@bitgo/utxo-lib';
 import GlobalCoinFactory from './v2/coinFactory';
@@ -19,19 +19,16 @@ export * as Errors from './errors';
 // Expose legacy "bitcoin" API (mostly HDNode)
 /** @deprecated */
 export * as bitcoin from './legacyBitcoin';
+
 /** @deprecated */
 export const sjcl = require('@bitgo/sjcl');
 
 export { Buffer } from 'buffer';
 
-// TODO - remove this once sdk-core is published. Express relies on this coming from bitgo
-export { CustomSigningFunction, SignedTransaction };
-
 export const Environments = _.cloneDeep(common.Environments);
-export { CoinConstructor, EnvironmentName, V1Network } from '@bitgo/sdk-core';
+export * from '@bitgo/sdk-core';
 export { GlobalCoinFactory };
 export * from './v2';
-export { tss };
 
 /**
  * Set the network, i.e. either "bitcoin" for production with real bitcoin, or
