+++ modules/bitgo/src/v2/index.ts
@@ -1,4 +1,3 @@
 export * from './coinFactory';
-export * from './types';
 import * as Coin from './coins';
 export { Coin };
