+++ modules/sdk-core/src/account-lib/index.ts
@@ -1,5 +1,9 @@
 export * from './baseCoin';
+// Deprecated
+export * as accountLibBaseCoin from './baseCoin';
 export * from './mpc';
 export * from './util/crypto';
+// Deprecated
+export * as acountLibCrypto from './util/crypto';
 export * from './util/ed25519KeyDeriver';
 export * from './staking';
