+++ package.json
@@ -57,6 +57,7 @@
     "eventsource": "2.0.2",
     "@types/react": "17.0.24",
     "@types/react-dom": "17.0.16",
+    "parse-path": "^5.0.0",
     "parse-url": "6.0.2"
   },
   "workspaces": [
